chore: move IBC utils and types to network package
