+++ packages/fast-usdc-deploy/src/upgrade-evm-dests.core.js
@@ -16,7 +16,8 @@ const { make } = AmountMath;
 /**
  * @import {CopyRecord} from '@endo/pass-style';
  * @import {ManifestBundleRef} from '@agoric/deploy-script-support/src/externalTypes.js';
- * @import {CosmosChainInfo, IBCConnectionInfo} from '@agoric/orchestration';
+ * @import {IBCConnectionInfo} from '@agoric/network/ibc';
+ * @import {CosmosChainInfo} from '@agoric/orchestration';
  * @import {BundleID} from '@agoric/swingset-vat';
  * @import {BootstrapManifest} from '@agoric/vats/src/core/lib-boot.js';
  * @import {FastUSDCCorePowers} from './start-fast-usdc.core.js';
