+++ packages/network/package.json
@@ -21,6 +21,7 @@
   "author": "Agoric",
   "license": "Apache-2.0",
   "dependencies": {
+    "@agoric/cosmic-proto": "workspace:*",
     "@agoric/internal": "workspace:*",
     "@agoric/store": "workspace:*",
     "@agoric/vat-data": "workspace:*",
@@ -41,7 +42,9 @@
     "c8": "^10.1.3"
   },
   "exports": {
-    ".": "./src/index.js"
+    ".": "./src/index.js",
+    "./ibc": "./src/ibc/index.js",
+    "./ibc/utils.js": "./src/ibc/utils.js"
   },
   "files": [
     "CHANGELOG.md",
