+++ packages/network/src/ibc/index.js
@@ -1,0 +1,1 @@
+export * from './types.ts';
