+++ packages/network/tsconfig.json
@@ -1,10 +1,9 @@
 {
   "extends": "../../tsconfig.json",
   "include": [
-    "*.js",
-    "scripts/**/*.js",
-    "src/**/*.js",
-    "test/**/*.js",
-    "tools/**/*.js",
+    "src",
+    "test",
+    "testing",
+    "tools"
   ],
 }
