+++ packages/orchestration/src/examples/auto-stake-it-tap-kit.js
@@ -8,7 +8,8 @@ import { CosmosChainAddressShape } from '../typeGuards.js';
 const trace = makeTracer('AutoStakeItTap');
 
 /**
- * @import {IBCChannelID, VTransferIBCEvent} from '@agoric/vats';
+ * @import {IBCChannelID} from '@agoric/network/ibc';
+ * @import {VTransferIBCEvent} from '@agoric/vats';
  * @import {VowTools} from '@agoric/vow';
  * @import {Zone} from '@agoric/zone';
  * @import {TargetApp} from '@agoric/vats/src/bridge-target.js';
