+++ packages/orchestration/src/examples/stake-ica.contract.js
@@ -21,7 +21,8 @@ const trace = makeTracer('StakeIca');
  * @import {ERemote, Remote} from '@agoric/internal';
  * @import {CosmosChainInfo, CosmosInterchainService, Denom, DenomDetail} from '@agoric/orchestration';
  * @import {ContractMeta, Invitation, ZCF, ZCFSeat} from '@agoric/zoe';
- * @import {IBCConnectionID, NameHub} from '@agoric/vats';
+ * @import {IBCConnectionID} from '@agoric/network/ibc';
+ * @import {NameHub} from '@agoric/vats';
  * @import {TimerService} from '@agoric/time';
  * @import {ResolvedContinuingOfferResult} from '../utils/zoe-tools.js';
  * @import {StorageNode} from '@agoric/internal/src/lib-chainStorage.js';
