+++ packages/orchestration/src/exos/cosmos-interchain-service.js
@@ -16,9 +16,8 @@ import { prepareICQConnectionKit } from './icq-connection-kit.js';
  * @import {Zone} from '@agoric/base-zone';
  * @import {Remote} from '@agoric/internal';
  * @import {Connection, Port, PortAllocator} from '@agoric/network';
- * @import {IBCConnectionID} from '@agoric/vats';
+ * @import {IBCConnectionID, RemoteIbcAddress} from '@agoric/network/ibc';
  * @import {MapStore, SetStore} from '@agoric/store';
- * @import {RemoteIbcAddress} from '@agoric/vats/tools/ibc-utils.js';
  * @import {Vow, VowTools} from '@agoric/vow';
  * @import {ICQConnection, IcaAccount, ICQConnectionKit, IcaAccountKit} from '../types.js';
  * @import {ICAChannelAddressOpts} from '../utils/address.js';
