+++ packages/orchestration/src/proposals/start-stakeOsmo.js
@@ -8,7 +8,7 @@ import { registerChainsAndAssets } from '../utils/chain-hub-helper.js';
 
 /**
  * @import {CosmosChainInfo, Denom, DenomDetail} from '@agoric/orchestration';
- * @import {IBCConnectionID} from '@agoric/vats';
+ * @import {IBCConnectionID} from '@agoric/network/ibc';
  * @import {StakeIcaSF} from '../examples/stake-ica.contract.js';
  * @import {start} from '../examples/stake-ica.contract.js';
  * @import {Installation} from '@agoric/zoe';
