+++ packages/orchestration/test/exos/chain-hub.test.ts
@@ -2,7 +2,8 @@ import '@agoric/swingset-liveslots/tools/prepare-test-env.js';
 import test from '@endo/ses-ava/prepare-endo.js';
 
 import { makeIssuerKit } from '@agoric/ertp';
-import { makeNameHubKit, type IBCChannelID } from '@agoric/vats';
+import type { IBCChannelID, IBCConnectionInfo } from '@agoric/network/ibc';
+import { makeNameHubKit } from '@agoric/vats';
 import { prepareSwingsetVowTools } from '@agoric/vow/vat.js';
 import { withAmountUtils } from '@agoric/zoe/tools/test-utils.js';
 import { E } from '@endo/far';
@@ -14,7 +15,6 @@ import {
 import type {
   Bech32Address,
   CosmosChainInfo,
-  IBCConnectionInfo,
 } from '../../src/cosmos-api.js';
 import { makeChainHub, registerAssets } from '../../src/exos/chain-hub.js';
 import knownChains from '../../src/fetched-chain-info.js';
