+++ packages/orchestration/test/facade-durability.test.ts
@@ -2,7 +2,8 @@ import { test as anyTest } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
 import { makeIssuerKit } from '@agoric/ertp';
 import { prepareSwingsetVowTools } from '@agoric/vow/vat.js';
-import type { CosmosChainInfo, IBCConnectionInfo } from '../src/cosmos-api.js';
+import type { IBCConnectionInfo } from '@agoric/network/ibc';
+import type { CosmosChainInfo } from '../src/cosmos-api.js';
 import fetchedChainInfo from '../src/fetched-chain-info.js'; // Refresh with scripts/refresh-chain-info.ts
 import type { Chain } from '../src/orchestration-api.js';
 import { denomHash } from '../src/utils/denomHash.js';
