+++ packages/orchestration/test/utils/address.test.ts
@@ -1,4 +1,4 @@
-import { decodeRemoteIbcAddress } from '@agoric/vats/tools/ibc-utils.js';
+import { decodeRemoteIbcAddress } from '@agoric/network/ibc/utils.js';
 import test from '@endo/ses-ava/prepare-endo.js';
 import type { Bech32Address } from '../../src/cosmos-api.ts';
 import type { CosmosChainAddress } from '../../src/orchestration-api.ts';
