+++ packages/orchestration/tools/ibc-mocks.ts
@@ -19,12 +19,8 @@ import {
   IBC_EVENT,
   VTRANSFER_IBC_EVENT,
 } from '@agoric/internal/src/action-types.js';
-import type {
-  IBCChannelID,
-  IBCEvent,
-  IBCPacket,
-  VTransferIBCEvent,
-} from '@agoric/vats';
+import type { IBCChannelID } from '@agoric/network/ibc';
+import type { IBCEvent, IBCPacket, VTransferIBCEvent } from '@agoric/vats';
 import { LOCALCHAIN_DEFAULT_ADDRESS } from '@agoric/vats/tools/fake-bridge.js';
 import { atob, btoa, decodeBase64, encodeBase64 } from '@endo/base64';
 import type { CosmosChainAddress, Denom } from '../src/orchestration-api.js';
