+++ packages/portfolio-deploy/src/chain-name-service.js
@@ -1,7 +1,8 @@
 import { denomHash } from '@agoric/orchestration/src/utils/denomHash.js';
 
 /**
- * @import {IBCChannelID, NameAdmin} from '@agoric/vats';
+ * @import {IBCChannelID} from '@agoric/network/ibc';
+ * @import {NameAdmin} from '@agoric/vats';
  * @import {ChainInfo, Denom} from '@agoric/orchestration';
  * @import {CosmosChainInfo} from '@agoric/orchestration';
  * @import {DenomDetail} from '@agoric/orchestration';
