+++ packages/portfolio-deploy/src/orch.start.js
@@ -7,7 +7,8 @@ import { fromExternalConfig } from './config-marshal.js';
  * @import { ManifestBundleRef } from '@agoric/deploy-script-support/src/externalTypes.js';
  * @import { Remote, ERemote } from '@agoric/internal';
  * @import { Marshaller, StorageNode } from '@agoric/internal/src/lib-chainStorage.js';
- * @import { ChainInfo, IBCConnectionInfo } from '@agoric/orchestration';
+ * @import { IBCConnectionInfo } from '@agoric/network/ibc';
+ * @import { ChainInfo } from '@agoric/orchestration';
  * @import { Board, NameHub } from '@agoric/vats';
  * @import { BootstrapManifest, BootstrapManifestPermit } from '@agoric/vats/src/core/lib-boot.js';
  * @import { Installation, IssuerKeywordRecord } from '@agoric/zoe';
